isolate delegate probation page components
